Demo_app/backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
facenet-pytorch>=2.5.3
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0